Simplify MATLAB architecture with proper naming and eliminate unnecessary layers
## Architecture Improvements
- Rename matlab_runner.m → matlab_interface.m (actual interface function)
- Remove matlab_runner.py (unnecessary MatlabSolver middle layer)
- Update matlab_interface.py to call matlab_interface.m directly

## Role Clarification
- Loader: problem_name → A,b,c,K (mat_loader.m, dat_loader.m)
- Runner: A,b,c,K → result (sedumi_runner.m, sdpt3_runner.m)
- Interface: problem_name → SolverResult (matlab_interface.m orchestrates loader + runner)
- Wrapper: Python wrapper for MATLAB interface (matlab_interface.py)

## Benefits
- Symmetrical architecture with python_interface.py
- Reduced coupling: new solvers require only 2 file changes (runner.m + config)
- Eliminated 3-layer abstraction (Python wrapper → Python solver → MATLAB script)
- Direct subprocess calls to matlab_interface.m
- Dynamic feval() preserved for flexible solver addition

## Technical Changes
- Direct MATLAB interface calls via subprocess
- Maintained dynamic runner function selection
- Preserved all existing functionality
- Compatible with existing benchmark runner
